requirements.txt
beautifulsoup4==4.14.3
Flask==3.1.3
google-genai==1.70.0
imageio-ffmpeg==0.6.0
openai==2.30.0
pandas==3.0.2
Pillow==12.2.0
python-dotenv==1.2.2
rapidocr-onnxruntime==1.4.4
requests==2.33.1
selenium==4.41.0
SQLAlchemy==2.0.49
streamlit==1.56.0
webdriver-manager==4.0.2
